agent/requirements.txt
anthropic
aiohttp
python-dotenv
